chore: refactor common logic
