+++ bouncer/shared/initialize_new_chains.ts
@@ -1,4 +1,5 @@
 import Web3 from 'web3';
+import { ApiPromise } from '@polkadot/api';
 import {
   Connection,
   NonceAccount,
@@ -12,10 +13,19 @@ import {
   getSolWhaleKeyPair,
   encodeSolAddress,
   getSolConnection,
+  observeEvent,
 } from '../shared/utils';
 import { signAndSendTxSol } from '../shared/send_sol';
 import { getSolanaVaultIdl, getKeyManagerAbi } from '../shared/contract_interfaces';
 import { signAndSendTxEvm } from '../shared/send_evm';
+import { submitGovernanceExtrinsic } from './cf_governance';
+
+export async function initializeArbitrumChain(chainflip: ApiPromise) {
+  console.log('Initializing Arbitrum');
+  const arbInitializationRequest = observeEvent('arbitrumVault:ChainInitialized', chainflip);
+  await submitGovernanceExtrinsic(chainflip.tx.arbitrumVault.initializeChain());
+  await arbInitializationRequest;
+}
 
 export async function initializeArbitrumContracts(
   arbClient: Web3,
